Changed: Extract tablecell-post-fixer.
